Move pipeline files into stack pkg
